+++ src/Exceptions/PresetNotAvailable.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Stolt\LeanPackage\Exceptions;
+
+class PresetNotAvailable extends \Exception
+{
+
+}