+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
+- New `--preset` option. Closes [#43](https://github.com/raphaelstolt/lean-package-validator/issues/43).
+
+### Added
+
 - New `--report-stale-export-ignores` option. Closes [#41](https://github.com/raphaelstolt/lean-package-validator/issues/41).
 
 ## [v3.1.1] - 2023-10-13